add restriction

+++ VisitorPattern/NutritionInfo.cs
@@ -4,6 +4,6 @@
     {
         public int HealthRating { get; set; }
         public int Calories { get; set; }
-        public int Protein { get; set; } 
+        public decimal Protein { get; set; } 
     }
 }
